% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.CoxBoost-package}
\alias{mlr3learners.CoxBoost}
\alias{mlr3learners.CoxBoost-package}
\title{mlr3learners.CoxBoost: Learners from the {CoxBoost} package for 'mlr3'}
\description{
Adds learners from the {CoxBoost} package to
    'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
